refactor(#35): mysql 패키지 분리
* mysql 관련 도메인 mysql 패키지로 묶어주기
* JpaConfig 설정 추가
