+++ backend/src/main/java/com/example/backend/auth/api/service/auth/AuthService.java
@@ -3,14 +3,12 @@ package com.example.backend.auth.api.service.auth;
 import com.example.backend.auth.api.controller.auth.response.AuthLoginResponse;
 import com.example.backend.auth.api.service.oauth.OAuthService;
 import com.example.backend.auth.api.service.oauth.response.OAuthResponse;
-import com.example.backend.domain.define.user.constant.UserPlatformType;
-import com.example.backend.domain.define.user.repository.UserRepository;
+import com.example.backend.domain.mysql.define.user.constant.UserPlatformType;
+import com.example.backend.domain.mysql.define.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
 
 @Slf4j
 @Service
