+++ backend/src/test/java/com/example/backend/auth/api/service/jwt/JwtServiceTest.java
@@ -1,18 +1,14 @@
 package com.example.backend.auth.api.service.jwt;
 
 import com.example.backend.auth.TestConfig;
-import com.example.backend.common.exception.ExceptionMessage;
-import com.example.backend.common.exception.jwt.JwtException;
-import com.example.backend.domain.define.user.User;
-import com.example.backend.domain.define.user.constant.UserRole;
-import com.example.backend.domain.define.user.repository.UserRepository;
+import com.example.backend.domain.mysql.define.user.User;
+import com.example.backend.domain.mysql.define.user.repository.UserRepository;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.ExpiredJwtException;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
 
 import java.util.Date;
 import java.util.HashMap;
@@ -20,7 +16,6 @@ import java.util.HashMap;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
 
 class JwtServiceTest extends TestConfig {
