+++ backend/src/test/java/com/example/backend/domain/define/user/repository/UserRepositoryTest.java
@@ -1,7 +1,8 @@
 package com.example.backend.domain.define.user.repository;
 
 import com.example.backend.auth.TestConfig;
-import com.example.backend.domain.define.user.User;
+import com.example.backend.domain.mysql.define.user.User;
+import com.example.backend.domain.mysql.define.user.repository.UserRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
